[REFACTOR-022] decompose: move ABO queries/mutations into AboInfoContent, EmailPrefsSection reads profile from cache
